checks full forms exceptions
