+++ packages/yoastseo/src/morphology/german/generateNounExceptionForms.js
@@ -1,46 +1,4 @@
-/**
- * Checks whether a stemmed word is on the exception list for which we have full forms.
- *
- * @param {Array} exceptionStems        The exception stems to check against.
- * @param {string} stemmedWordToCheck   The stem to check.
- *
- * @returns {string[]} The created word forms.
- */
-const checkStemsFromExceptionList = function( exceptionStems, stemmedWordToCheck ) {
-	for ( let i = 0; i < exceptionStems.length; i++ ) {
-		const currentStemDataSet = exceptionStems[ i ];
-
-		const stemPairToCheck = currentStemDataSet[ 0 ];
-
-		for ( let j = 0; j < stemPairToCheck.length; j++ ) {
-			const exceptionStemMatched = stemmedWordToCheck.endsWith( stemPairToCheck[ j ] );
-
-			// Check if the stemmed word ends in one of the stems of the exception list.
-			if ( exceptionStemMatched === true ) {
-				// "Haupt".length = "Hauptstadt".length - "stadt".length
-				const precedingLength = stemmedWordToCheck.length - stemPairToCheck[ j ].length;
-				const precedingLexicalMaterial = stemmedWordToCheck.slice( 0, precedingLength );
-				/*
-			 	 * If the word is a compound, removing the final stem will result in some lexical material to
-			 	 * be left over at the beginning of the word. For example, removing "stadt" from "Hauptstadt"
-			 	 * leaves "Haupt". This lexical material is the base for the word forms that need to be created
-			 	 * (e.g., "Hauptstädte"). The preceding lexical material must include at least 2 characters in order to be a valid compound element.
-			 	 */
-				if ( precedingLexicalMaterial.length > 1 ) {
-					const stemsToReturn = currentStemDataSet[ 1 ];
-					return stemsToReturn.map( currentStem => precedingLexicalMaterial.concat( currentStem ) );
-				}
-				/*
-				 * Return all possible stems since apparently the word that's being checked is equal to the stem on the
-				 * exception list that's being checked.
-				 */
-				return currentStemDataSet[ 1 ];
-			}
-		}
-	}
-
-	return [];
-};
+import checkExceptionsWithFullForms from "../../morphology/morphoHelpers/checkExceptionsWithFullForms";
 
 /**
  * Checks whether a stemmed word has an ending for which we can predict possible suffix forms.
@@ -81,7 +39,7 @@ const checkStemsWithPredictableSuffixes = function( exceptionCategory, stemmedWo
  */
 export function generateNounExceptionForms( morphologyDataNouns, stemmedWordToCheck ) {
 	// Check exceptions with full forms.
-	let exceptions = checkStemsFromExceptionList( morphologyDataNouns.exceptionStemsWithFullForms, stemmedWordToCheck );
+	let exceptions = checkExceptionsWithFullForms( morphologyDataNouns.exceptionStemsWithFullForms, stemmedWordToCheck );
 
 	if ( exceptions.length > 0 ) {
 		return exceptions;
